// FE.Core.FW/src/apis/models/RoleModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type RoleModel = {
  id?: string;
  code?: string | null;
  name?: string | null;
  description?: string | null;
  recordCloneId?: string | null;
  lastModifiedOnDate?: string;
  createdOnDate?: string;
  isDefault?: boolean;
  accessDataHeaderQuater?: string | null;
};